Minor updates
- added parameter timestamp (false by default) to input file. If true it
prints in the output timestamp in miliseconds.
- added new mode saturate, which gets a non-saturated packing and add
shapes to it up to saturation

+++ rsa3d/Config.cpp
@@ -67,7 +67,7 @@ int Config::getInt(const std::string & field) const {
 
 unsigned long Config::getUnsignedLong(const std::string & field) const {
     auto str = this->getString(field);
-    if (std::stoi(str) < 0)
+    if (std::stol(str) < 0)
         throw std::invalid_argument("unsigned long field negative");
     return std::stoul(str);
 }
